Optimize Global Filter saving on disk performance
Initially we were doing polling to save the filter periodically,
which needed thread-safety and hence slowed down data insertion.

Now we're flushing the filter when Cassandra is flushing SSTable
filters. However, even then, we're constantly checking the hash
of the filter before saving to see if it really changed and only
then flush the filter.